Create models package and create a module for each model
